+ reorganising the project

+++ src/Integers.cs
@@ -1,12 +1,14 @@
+using System.Numerics;
+
 namespace PMath
 {
     public static class Integers
     {
-        public static int EuclidGCD(int a, int b)
+        public static BigInteger EuclidGCD(BigInteger a, BigInteger b)
         {
             while (b > 0)
             {
-                int remainder = a % b;
+                BigInteger remainder = a % b;
                 a = b;
                 b = remainder;
             }
